+++ eras/mary/impl/src/Cardano/Ledger/Mary/Core.hs
@@ -4,12 +4,12 @@ module Cardano.Ledger.Mary.Core (
   mintedAssetsTxBodyF,
   burnedAssetsTxBodyF,
   mintPoliciesTxBodyF,
-  module Cardano.Ledger.Mary.Mint,
+  module Cardano.Ledger.Mary.Forging,
   module Cardano.Ledger.Allegra.Core,
 ) where
 
 import Cardano.Ledger.Allegra.Core
-import Cardano.Ledger.Mary.Mint
+import Cardano.Ledger.Mary.Forging
 import Cardano.Ledger.Mary.Tx ()
 import Cardano.Ledger.Mary.TxBody (
   MaryEraTxBody (..),